requirements.txt
asn1crypto==0.24.0
awesome-slugify==1.6.5
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
cffi==1.12.1
cryptography==2.5
future==0.17.1
peewee==3.8.2
pycparser==2.19
python-decouple==3.1
python-telegram-bot==11.1.0
regex==2019.2.21
six==1.12.0
soupsieve==1.8
Unidecode==0.4.21
